requirements/requirements.txt
tensorflow
torch
torchvision
torchaudio
numpy
matplotlib
pandas
scipy
DateTime
openpyxl
requests
pathlib
argparse
tqdm
scikit-learn
psutil
pyyaml
lime
seaborn